refactor: 💡 restructure salesman screen company details  folder
